+++ tests/test_execution.py
@@ -1385,7 +1385,9 @@ class BetdaqExecutionTest(unittest.TestCase):
         mock_order_package.__len__.return_value = 1
         mock_order_package.__iter__ = mock.Mock(return_value=iter([mock_order]))
         mock_order_package.info = {}
-        mock_instruction_report = [{"return_code": None, "customer_reference": 123}]
+        mock_instruction_report = [
+            {"return_code": None, "customer_reference": 123, "status": "Unmatched"}
+        ]
         mock___execution_helper.return_value = mock_instruction_report
 
         self.execution.execute_place(mock_order_package, mock_session)
@@ -1417,7 +1419,9 @@ class BetdaqExecutionTest(unittest.TestCase):
             return_value=iter([mock_order, mock_order_two])
         )
         mock_order_package.info = {}
-        mock_instruction_report = [{"return_code": None, "customer_reference": 123}]
+        mock_instruction_report = [
+            {"return_code": None, "customer_reference": 123, "status": "Unmatched"}
+        ]
         mock___execution_helper.return_value = mock_instruction_report
 
         self.execution.execute_place(mock_order_package, mock_session)
@@ -1444,7 +1448,9 @@ class BetdaqExecutionTest(unittest.TestCase):
         mock_order_package.__len__.return_value = 1
         mock_order_package.__iter__ = mock.Mock(return_value=iter([mock_order]))
         mock_order_package.info = {}
-        mock_instruction_report = [{"return_code": 123, "customer_reference": 123}]
+        mock_instruction_report = [
+            {"return_code": 123, "customer_reference": 123, "status": "Unmatched"}
+        ]
         mock___execution_helper.return_value = mock_instruction_report
 
         self.execution.execute_place(mock_order_package, mock_session)
@@ -1477,6 +1483,33 @@ class BetdaqExecutionTest(unittest.TestCase):
         mock__order_logger.assert_not_called()
         mock_order.executable.assert_not_called()
         mock_order.execution_complete.assert_called()
+        mock_order.trade.__enter__.assert_called_with()
+        mock_order.trade.__exit__.assert_called_with(None, None, None)
+        mock___execution_helper.assert_called_with(
+            self.execution.place, mock_order_package
+        )
+
+    @mock.patch("flumine.execution.betdaqexecution.BetdaqExecution._order_logger")
+    @mock.patch("flumine.execution.betdaqexecution.BetdaqExecution._execution_helper")
+    def test_execute_place_async(self, mock___execution_helper, mock__order_logger):
+        mock_session = mock.Mock()
+        mock_order = mock.Mock(id="123")
+        mock_order.trade.__enter__ = mock.Mock()
+        mock_order.trade.__exit__ = mock.Mock()
+        mock_order_package = mock.MagicMock()
+        mock_order_package.__len__.return_value = 1
+        mock_order_package.__iter__ = mock.Mock(return_value=iter([mock_order]))
+        mock_order_package.info = {}
+        mock_instruction_report = [
+            {"return_code": None, "customer_reference": 123, "status": None}
+        ]
+        mock___execution_helper.return_value = mock_instruction_report
+
+        self.execution.execute_place(mock_order_package, mock_session)
+        mock__order_logger.assert_called_with(
+            mock_order, mock_instruction_report[0], mock_order_package.package_type
+        )
+        mock_order.executable.assert_not_called()
         mock_order.trade.__enter__.assert_called_with()
         mock_order.trade.__exit__.assert_called_with(None, None, None)
         mock___execution_helper.assert_called_with(
@@ -1767,10 +1800,27 @@ class BetdaqExecutionTest(unittest.TestCase):
     @mock.patch("flumine.execution.betdaqexecution.OrderEvent")
     def test__order_logger_place(self, mock_order_event):
         mock_order = mock.Mock(async_=False)
-        mock_instruction_report = {"return_code": None, "order_id": 123}
+        mock_instruction_report = {
+            "return_code": None,
+            "order_id": 123,
+            "status": "Unmatched",
+        }
         self.execution._order_logger(
             mock_order, mock_instruction_report, OrderPackageType.PLACE
         )
         self.assertEqual(mock_order.bet_id, 123)
         mock_order.responses.placed.assert_called_with(mock_instruction_report, dt=True)
+        self.mock_flumine.log_control.assert_called_with(mock_order_event(mock_order))
+
+    @mock.patch("flumine.execution.betdaqexecution.OrderEvent")
+    def test__order_logger_place_async(self, mock_order_event):
+        mock_order = mock.Mock(async_=False)
+        mock_instruction_report = {"return_code": None, "order_id": 123, "status": None}
+        self.execution._order_logger(
+            mock_order, mock_instruction_report, OrderPackageType.PLACE
+        )
+        self.assertEqual(mock_order.bet_id, 123)
+        mock_order.responses.placed.assert_called_with(
+            mock_instruction_report, dt=False
+        )
         self.mock_flumine.log_control.assert_called_with(mock_order_event(mock_order))